Controller Tests Fix
