+++ src/test/java/uk/gov/hmcts/reform/finrem/caseorchestration/controllers/ContestedDraftOrderNotApprovedControllerTest.java
@@ -1,9 +1,11 @@
 package uk.gov.hmcts.reform.finrem.caseorchestration.controllers;
 
 import org.junit.Test;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.webmvc.test.autoconfigure.WebMvcTest;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
+import org.springframework.test.web.servlet.MockMvc;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 import uk.gov.hmcts.reform.finrem.caseorchestration.helper.DocumentHelper;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.ContestedDraftOrderNotApprovedService;
@@ -36,6 +38,8 @@ public class ContestedDraftOrderNotApprovedControllerTest extends BaseController
     private IdamService idamService;
     @MockitoBean
     private DocumentHelper documentHelper;
+    @Autowired
+    protected MockMvc mvc;
 
     private static final String SUBMIT_REFUSAL_REASON_URL = "/case-orchestration/contested-application-send-refusal";
