+++ src/test/java/uk/gov/hmcts/reform/finrem/caseorchestration/controllers/FeeLookupControllerTest.java
@@ -2,9 +2,11 @@ package uk.gov.hmcts.reform.finrem.caseorchestration.controllers;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.Test;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.webmvc.test.autoconfigure.WebMvcTest;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
+import org.springframework.test.web.servlet.MockMvc;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 import uk.gov.hmcts.reform.finrem.caseorchestration.error.GlobalExceptionHandler;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.ApplicationType;
@@ -40,6 +42,8 @@ public class FeeLookupControllerTest extends BaseControllerTest {
     private FeeService feeService;
     @MockitoBean
     private CaseDataService caseDataService;
+    @Autowired
+    protected MockMvc mvc;
 
     private void doFeeLookupSetUp(ApplicationType applicationType) throws IOException, URISyntaxException {
         String fileName = applicationType == CONSENTED
