+++ src/test/java/uk/gov/hmcts/reform/finrem/caseorchestration/controllers/GeneralOrderControllerTest.java
@@ -2,9 +2,11 @@ package uk.gov.hmcts.reform.finrem.caseorchestration.controllers;
 
 import org.junit.Test;
 import org.mockito.stubbing.OngoingStubbing;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.webmvc.test.autoconfigure.WebMvcTest;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
+import org.springframework.test.web.servlet.MockMvc;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.GeneralOrderService;
 
@@ -27,6 +29,9 @@ import static uk.gov.hmcts.reform.finrem.caseorchestration.TestSetUpUtils.feignE
 
 @WebMvcTest(GeneralOrderController.class)
 public class GeneralOrderControllerTest extends BaseControllerTest {
+
+    @Autowired
+    protected MockMvc mvc;
 
     @MockitoBean
     private GeneralOrderService documentService;
