+++ src/test/java/uk/gov/hmcts/reform/finrem/caseorchestration/controllers/PBAPaymentControllerTest.java
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.webmvc.test.autoconfigure.WebMvcTest;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
+import org.springframework.test.web.servlet.MockMvc;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.organisation.OrganisationContactInformation;
 import uk.gov.hmcts.reform.finrem.caseorchestration.model.organisation.OrganisationsResponse;
@@ -61,6 +62,8 @@ public class PBAPaymentControllerTest extends BaseControllerTest {
     private PrdOrganisationService prdOrganisationService;
     @MockitoBean
     private MiamLegacyExemptionsService miamLegacyExemptionsService;
+    @Autowired
+    protected MockMvc mvc;
 
     @Autowired
     private ObjectMapper objectMapper;
