+++ src/test/java/uk/gov/hmcts/reform/finrem/caseorchestration/controllers/PaymentConfirmationControllerTest.java
@@ -3,9 +3,11 @@ package uk.gov.hmcts.reform.finrem.caseorchestration.controllers;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.Before;
 import org.junit.Test;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.webmvc.test.autoconfigure.WebMvcTest;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
+import org.springframework.test.web.servlet.MockMvc;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 import uk.gov.hmcts.reform.finrem.caseorchestration.error.GlobalExceptionHandler;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.CaseDataService;
@@ -37,6 +39,8 @@ public class PaymentConfirmationControllerTest extends BaseControllerTest {
     private PaymentConfirmationService paymentConfirmationService;
     @MockitoBean
     private CaseDataService caseDataService;
+    @Autowired
+    protected MockMvc mvc;
 
     @Before
     public void setUp() {
